add scheduler api module
