+++ seatunnel-server/seatunnel-app/src/main/java/org/apache/seatunnel/app/plugin/SchedulerPluginHelper.java
@@ -1,0 +1,4 @@
+package org.apache.seatunnel.app.plugin;
+
+public class SchedulerPluginHelper {
+}
