+++ seatunnel-server/seatunnel-scheduler/seatunnel-scheduler-api/src/main/java/org/apache/seatunnel/scheduler/api/dto/ComplementDataDto.java
@@ -15,15 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.spi.scheduler.dto;
+package org.apache.seatunnel.scheduler.api.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
-@Builder
+import java.util.Date;
+
 @Data
-public class InstanceListDto {
-    private String name;
-    private int pageNo;
-    private int pageSize;
+@Builder
+public class ComplementDataDto {
+    private Date startTime;
+    private Date endTime;
+    private Integer parallelismNum;
 }
