+++ seatunnel-server/seatunnel-scheduler/seatunnel-scheduler-api/src/main/java/org/apache/seatunnel/scheduler/api/dto/ExecuteDto.java
@@ -15,19 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.spi.scheduler.dto;
+package org.apache.seatunnel.scheduler.api.dto;
+
+import org.apache.seatunnel.spi.scheduler.enums.ExecuteTypeEnum;
 
 import lombok.Builder;
 import lombok.Data;
 
-import java.util.Date;
-
 @Data
 @Builder
-public class SchedulerConfigDto {
-    private int retryTimes;
-    private int retryInterval;
-    private Date startTime;
-    private Date endTime;
-    private String triggerExpression;
+public class ExecuteDto {
+    private JobDto jobDto;
+    private ExecuteTypeEnum executeTypeEnum;
+    private ComplementDataDto complementDataDto;
+
 }
