+++ seatunnel-server/seatunnel-scheduler/seatunnel-scheduler-api/src/main/java/org/apache/seatunnel/scheduler/api/dto/JobDto.java
@@ -15,20 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.spi.scheduler.dto;
+package org.apache.seatunnel.scheduler.api.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
-import java.util.Date;
+import java.util.Map;
 
-@Builder
 @Data
-public class JobSimpleInfoDto {
-    private long jobId;
-    private String jobStatus;
-    private String creatorName;
-    private String menderName;
-    private Date createTime;
-    private Date updateTime;
+@Builder
+public class JobDto {
+    private Long jobId;
+    private String jobName;
+    private String executorScript;
+    private String jobContent;
+    private Integer operatorId;
+    private Map<String, Object> params;
+    private SchedulerConfigDto schedulerConfigDto;
 }
