+++ seatunnel-server/seatunnel-scheduler/seatunnel-scheduler-api/src/main/java/org/apache/seatunnel/scheduler/api/dto/JobListDto.java
@@ -15,18 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.spi.scheduler.dto;
-
-import org.apache.seatunnel.spi.scheduler.enums.ExecuteTypeEnum;
+package org.apache.seatunnel.scheduler.api.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-public class ExecuteDto {
-    private JobDto jobDto;
-    private ExecuteTypeEnum executeTypeEnum;
-    private ComplementDataDto complementDataDto;
-
+public class JobListDto {
+    private String name;
+    private Integer pageNo;
+    private Integer pageSize;
 }
